# Faster `dropdims`